Automatically parse and track library dependencies
This commit adds the posibility of only configuring and building the
given libraries, provided by the BOOST_CMAKE_LIBRARIES variable:

    $ cmake .. -DBOOST_CMAKE_LIBRARIES="thread filesystem chrono"

The root CMakeLists.txt file of the project scans the source tree to
ensure only the given libraries and their dependencies are configured.
This reduces A LOT cmake configuration time and buildtree size.